Corrected proof value encoding
