+++ src/utils/mod.rs
@@ -6,9 +6,6 @@ pub mod base58;
 #[cfg(feature = "w3c")]
 pub mod base64;
 
-#[cfg(feature = "w3c")]
-pub mod encoded_object;
-
 pub mod hash;
 
 pub mod query;
